chore(ssot): phases 5-7 dead-code, formatter consolidation, canonical enforcement
Phase 5: location SSOT
- delete locationLabels.ts zombie formatter from apps/web
- consolidate display/sanitization into shared/src/location/location.format.ts
- update normalizeModerationAd and locationService to use shared formatters
- add ssot-validator guard preventing local locationLabels re-creation

Phase 6: admin catalog UI audit
- confirm CatalogPageTemplate already centralises tab layout, no unsafe merge
- SparePartsTab: migrate manual useState+useCatalogQueryStateSync to canonical
  useCatalogTabState (removes 4 duplicated state declarations)

Phase 7: secondary structural duplications
- ActiveSubscriptionCard: remove local formatPlanName, import from @esparex/shared
- extract formatCoordinateLabel from geoUtils.ts into shared/src/utils/geoDisplay.ts
  (geo-math vs display-formatter separation, keeps geoUtils in baseline threshold)
- ssot-validator Rule B: upgrade val.warning to val.error for canonical collisions
  (zero violations confirmed; future re-declarations now fail repo:gate)
